# Go_Server/docs/swagger.yaml
definitions:
  define.AddFoodRequest:
    properties:
      foodicon:
        type: string
      foodname:
        type: string
      foodprocedure:
        type: string
      price:
        type: number
      remarks:
        type: string
      user_id:
        type: integer
      video:
        type: string
    type: object
...
  define.LoginPassWordRequest:
    properties:
      code:
        description: 登录验证码
        type: string
      password:
        description: 登录密码
        type: string
      username:
        description: 登录用户名
        type: string
    type: object
  define.UpdateFoodRequest:
    properties:
      foodicon:
        type: string
      foodname:
        type: string
      foodprocedure:
        type: string
      id:
        type: integer
      price:
        type: number
      remarks:
        type: string
      user_id:
        type: integer
      video:
        type: string
    type: object
  define.UpdateInfoType:
    properties:
      avatar:
        type: string
      id:
        type: integer
      sex:
        type: string
      username:
        type: string
    type: object
...
info:
  contact:
    name: Vulnerabilities_Server
    url: https://github.com/A7cc/Vulnerabilities_Server
  description: 这是一个用Golang写的食谱菜单管理系统靶场，该系统中存在多个安全漏洞，需要我们去探索和发现。\n\n该项目旨在帮助安全研究人员和爱好者了解和掌握关于Golang系统的渗透测试和代码审计知识。
  title: 食谱菜单管理系统靶场
  version: "1.0"
paths:
  /auth/login:
    post:
      parameters:
      - description: 登录信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/define.LoginPassWordRequest'
      responses: {}
      summary: 处理用户登录
      tags:
      - 公共方法
...
  /food/add:
    post:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 接收添加菜品表单数据
        in: body
        name: AddFoodRequest
        required: true
        schema:
          $ref: '#/definitions/define.AddFoodRequest'
      responses: {}
      summary: 添加菜品
      tags:
      - 鉴权接口-菜品相关方法
  /food/delete/{id}:
    delete:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 菜品ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 删除菜品
      tags:
      - 鉴权接口-菜品相关方法
  /food/detail:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取菜品ID
        in: query
        name: id
        required: true
        type: integer
      responses: {}
      summary: 获取菜品详情信息
      tags:
      - 鉴权接口-菜品相关方法
  /food/get:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: keyword
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: size
        type: integer
      - in: query
        name: status
        type: integer
      responses: {}
      summary: 获取菜品列表
      tags:
      - 鉴权接口-菜品相关方法
  /food/update:
    put:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新菜品信息
        in: body
        name: UpdateFoodRequest
        required: true
        schema:
          $ref: '#/definitions/define.UpdateFoodRequest'
      responses: {}
      summary: 更新菜品信息
      tags:
      - 鉴权接口-菜品相关方法
  /food/upfoodicon:
    post:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新菜品ICON
        in: formData
        name: file
        required: true
        type: file
      responses: {}
      summary: 更新菜品ICON
      tags:
      - 鉴权接口-菜品相关方法
  /food/upfoodvideo:
    post:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新菜品视频
        in: formData
        name: file
        required: true
        type: file
      responses: {}
      summary: 更新菜品Video
      tags:
      - 鉴权接口-菜品相关方法
  /home/get:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 获取系统数据信息
      tags:
      - 鉴权接口-首页设置方法
  /home/getsentence:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取每日金句的url
        in: query
        name: url
        required: true
        type: string
      responses: {}
      summary: 获取每日金句
      tags:
      - 鉴权接口-首页设置方法
  /home/updateInfo:
    put:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户自己更新个人信息参数
        in: body
        name: UpdateInfoType
        required: true
        schema:
          $ref: '#/definitions/define.UpdateInfoType'
      responses: {}
      summary: 用户自己更新个人信息
      tags:
      - 鉴权接口-首页设置方法
  /home/updatePwd:
    put:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户自己更新密码和UID
        in: body
        name: newpwdinfo
        required: true
        schema:
          type: string
      responses: {}
      summary: 更改个人密码
      tags:
      - 鉴权接口-首页设置方法
  /home/upuseravatar:
    post:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新头像
        in: formData
        name: file
        required: true
        type: file
      responses: {}
      summary: 更新头像
      tags:
      - 鉴权接口-首页设置方法
...
swagger: "2.0"